combine general form and det equations
