+++ src/symmetries/objects/determining_equations.py
@@ -4,6 +4,7 @@ from copy import deepcopy
 import sympy
 import numpy as np
 from symmetries.utils.algebra import key_ordering, str_to_dict, is_zero
+from symmetries.utils.symbolic import get_symbolic_terms, parse_variables
 from .system_of_equations import SystemOfEquations
 from .system import System
 
@@ -19,7 +20,7 @@ class DeterminingEquations(SystemOfEquations):
                  constants: list,
                  ):
 
-        self.model_info = system
+        self.model = system
         self.rules_array = rules_array
 
         self.determining_equations_extended = ""
@@ -27,14 +28,19 @@ class DeterminingEquations(SystemOfEquations):
 
         super().__init__(independent_variables, dependent_variables, constants)
 
+        self.general_form = self.obtain_general_form()
+        self.parsed_variables = parse_variables(
+            independent_variables, dependent_variables)
+        self.deleted = {}
+
     def get_group_operator(self):
         """Given a differential equation F, gives the Lie operator acting over F.
         """
-        variables = self.all_variables + self.model_info.derivatives_subscript_notation
+        variables = self.all_variables + self.model.derivatives_subscript_notation
         l_f = 0
-        for var, inft in zip(variables, self.model_info.infinitesimals):
+        for var, inft in zip(variables, self.model.infinitesimals):
             l_f += inft * \
-                sympy.Derivative(self.model_info.differential_equation, var)
+                sympy.Derivative(self.model.differential_equation, var)
         self.determining_equations_extended = sympy.simplify(l_f)
 
     def variable_relabeling(self):
@@ -57,9 +63,9 @@ class DeterminingEquations(SystemOfEquations):
         """
 
         new_labeling = deepcopy(
-            self.model_info.dependent_variables_partial_derivatives)
+            self.model.dependent_variables_partial_derivatives)
         previous_labeling = deepcopy(
-            self.model_info.derivatives_subscript_notation)
+            self.model.derivatives_subscript_notation)
 
         new_labeling.reverse()
         previous_labeling.reverse()
@@ -218,3 +224,131 @@ class DeterminingEquations(SystemOfEquations):
         for idx in zero_terms:
             simplify_det_eqns[idx] = [zero_terms[idx]]
         return simplify_det_eqns
+
+    def obtain_general_form(self):
+        """Proposes a general form of solutions for all infinitesimals as a function of
+        all the independent and dependent variables accordingly.
+
+        Returns
+        -------
+        dictionary
+            Keys are the infinitesimals eta and eta and values are list of relevant variables.
+        """
+        general_form = {}
+        infinitesimals = self.model.infinitesimals_ind + self.model.infinitesimals_dep
+
+        for inft in infinitesimals:
+            l = re.split(r'\W+', str(inft), len(self.all_variables)+1)
+            if l[0] not in general_form:
+                general_form[l[0]] = {l[1]: deepcopy(self.all_variables)}
+            else:
+                general_form[l[0]][l[1]] = deepcopy(self.all_variables)
+
+        return general_form
+
+    def find_first_derivative_equals_0(self):
+        """Finds equations in the determining equations that contain a single term that is a first
+        derivative of a given variable and stores it in a dictionary that contains deleted
+        dependencies called deleted."""
+        for v in self.determining_equations.values():
+            if len(v) == 1:
+                item = v[0]
+                if sum(item['derivatives']) == 1:
+                    var = [v for v, order in zip(
+                        self.all_variables, item['derivatives']) if order][0]
+                    if not ((item['variable'] in self.deleted) and (
+                            var in self.deleted[item['variable']])):
+                        self.general_form[item['variable'][:-1]
+                                          ][item['variable'][-1]].remove(var)
+                        if item['variable'] not in self.deleted:
+                            self.deleted[item['variable']] = [var]
+                            print('deleting', item['variable'], var)
+                        else:
+                            self.deleted[item['variable']].append(var)
+                            print('deleting', item['variable'], var)
+
+    def find_deleted_items_in_equations(self):
+        """Searches in the determining equations for equations that contain a term that is an
+        already identified first derivative equals 0 or higher order or cross derivative of the
+        same."""
+        for k, eq in self.determining_equations.items():
+            if len(eq) > 1:
+                values = deepcopy(eq)
+                for item in values:
+                    if item['variable'] in self.deleted:
+                        variables = [v for v, order in zip(
+                            self.all_variables, item['derivatives']) if order]
+                        for variable in variables:
+                            if (variable in self.deleted[item['variable']]) and (item in eq):
+                                print('found', variable, 'in',
+                                      item['variable'], 'eq', k)
+                                eq.remove(item)
+
+    def delete_second_derivatives(self):
+        """Similarly to find_deleted_items_in_equations, the method iterates over the system of
+        equations, finds equations containing single higher order derivatives and deletes them
+        from the system.
+        Note: this is a separate method because it modifies the iterator."""
+        det_eqns = deepcopy(self.determining_equations)
+        for k, eq in det_eqns.items():
+            if len(eq) == 1:
+                item = eq[0]  # there is only a single term in the equation
+                if item['variable'] in self.deleted:
+                    variables = [v for v, order in zip(
+                        self.all_variables, item['derivatives']) if order > 1]
+                    for variable in variables:
+                        if variable in self.deleted[item['variable']]:
+                            print('found derivative of', variable,
+                                  'in', item['variable'], 'eq', k)
+                            del self.determining_equations[k]
+
+    def simplify_iteratively(self):
+        """Iterative method, perform the three steps until determining equations does not change.
+        """
+        while True:
+            check_against = deepcopy(self.determining_equations)
+            self.simplify_redundant_equations()
+            self.find_first_derivative_equals_0()
+            self.find_deleted_items_in_equations()
+            self.delete_second_derivatives()
+
+            if check_against == self.determining_equations:
+                break
+
+    def print_determining_equations(self):
+        """Prints determining equations as symbolic using sympy Matrix."""
+        return self._print_symbolic_equations(
+            self.determining_equations, 'determining')
+
+    def print_general_form(self):
+        """Prints general form as symbolic using sympy Matrix."""
+        print('already deleted:', self.deleted)
+        print('general form:')
+        return self._print_symbolic_equations(
+            self.general_form, 'general')
+
+    def _print_symbolic_equations(self, equations, type):
+        """Gives the symbolic version of remaining determining equation.
+
+        Parameters
+        ----------
+        det_eqn : dict
+            dictionary will all the determining equations.
+        """
+        matrix = sympy.Matrix([[]])
+        i = 0
+        for variable, eqns in equations.items():
+            row = None
+            if type == 'determining':
+                i += 1
+                row = sympy.Matrix([[i, sympy.Eq(get_symbolic_terms(
+                    eqns, self.parsed_variables, self.constants, self.all_variables), 0)]])
+                matrix = matrix.row_insert(i, row)
+
+            elif type == 'general':
+                for var, dependencies in eqns.items():
+                    row = sympy.Matrix(
+                        [f"{variable}({','.join([str(dep) for dep in dependencies])})^{var} "])
+                    matrix = matrix.row_insert(i, row)
+
+        return matrix
